requirements.txt
fastapi
uvicorn[standard]
websockets==12.0
python-dotenv
requests
sounddevice
numpy